Ex 13.8

+++ controllers/blogs.js
@@ -1,27 +1,26 @@
-const Blog = require("../models/blog");
+const { Blog } = require("../models/index.js");
 
-const router = require("express").Router();
+const blogsRouter = require("express").Router();
 
-router.get("/", async (req, res) => {
+blogsRouter.get("/", async (req, res) => {
   const blogs = await Blog.findAll();
   res.json(blogs);
-  res.status(400).json({ error: error });
 });
 
-router.post("/", async (req, res) => {
+blogsRouter.post("/", async (req, res) => {
   const blog = req.body;
   const createdBlog = await Blog.create(blog);
   res.status(201).json(createdBlog);
 });
 
-router.put("/:id", async (req, res, next) => {
+blogsRouter.put("/:id", async (req, res, next) => {
   const blogId = req.params.id;
   const newLikes = req.body.likes;
 
   if (!Number.isInteger(Number(blogId))) {
-    const error = new Error("Invalid ID format");
-    error.name = "BadRequestError";
-    return next(error); // Pass the error to the error handler
+    const badRequestError = new Error("Invalid ID format");
+    badRequestError.name = "BadRequestError";
+    return next(badRequestError); // Pass the error to the error handler
   }
 
   const blog = await Blog.findByPk(blogId);
@@ -36,11 +35,19 @@ router.put("/:id", async (req, res, next) => {
 
   const updatedBlog = await blog.save();
 
+  console.log(updatedBlog);
+
   res.status(200).json({ likes: newLikes });
 });
 
-router.delete("/:id", async (req, res) => {
+blogsRouter.delete("/:id", async (next, req, res) => {
   const blogId = req.params.id;
+
+  if (!Number.isInteger(Number(blogId))) {
+    const badRequestError = new Error("Invalid ID format");
+    badRequestError.name = "BadRequestError";
+    return next(badRequestError); // Pass the error to the error handler
+  }
 
   const result = await Blog.destroy({
     where: {
@@ -51,4 +58,4 @@ router.delete("/:id", async (req, res) => {
   res.status(204).json({ result });
 });
 
-module.exports = router;
+module.exports = blogsRouter;
